feat:Added templates for page creation
