+++ gatsby-node.ts
@@ -1,9 +1,10 @@
-const path = require("path")
+const path = require("path");
 
 module.exports.createPages = async ({ graphql, actions }) => {
-  const { createPage } = actions
-  const blogPostTemplate = path.resolve("src/templates/blog-post.tsx")
-  const pageTemplate = path.resolve("src/templates/page.tsx")
+  const { createPage } = actions;
+  const blogPostTemplate = path.resolve("src/templates/blog-post.tsx");
+  const pageTemplate = path.resolve("src/templates/page.tsx");
+  const blogPageTemplate = path.resolve("src/templates/blog-page.tsx");
   const blogPostQuery = await graphql(`
     query {
       allContentstackBlogPost {
@@ -13,7 +14,7 @@ module.exports.createPages = async ({ graphql, actions }) => {
         }
       }
     }
-  `)
+  `);
 
   const pageQuery = await graphql(`
     query {
@@ -24,33 +25,52 @@ module.exports.createPages = async ({ graphql, actions }) => {
         }
       }
     }
-  `)
+  `);
 
-  const createBlogPostTemplate = (route, comp, title) => {
+  const createBlogPostTemplate = (route, componentToRender, title) => {
     createPage({
       path: `${route}`,
-      component: comp,
+      component: componentToRender,
       context: {
         title: title,
       },
-    })
-  }
+    });
+  };
 
-  const createPageTemplate = (route, comp, url) => {
+  const createPageTemplate = (route, componentToRender, url) => {
     createPage({
       path: `${route}`,
-      component: comp,
+      component: componentToRender,
       context: {
         url: url,
       },
-    })
-  }
+    });
+  };
+
+  const createBlogPageTemplate = (route, componentToRender, title, data) => {
+    createPage({
+      path: `${route}`,
+      component: componentToRender,
+      context: {
+        title: title,
+        result: data,
+      },
+    });
+  };
+
   blogPostQuery.data.allContentstackBlogPost.nodes.forEach(node => {
-    createBlogPostTemplate(node.url, blogPostTemplate, node.title)
-  })
+    createBlogPostTemplate(node.url, blogPostTemplate, node.title);
+  });
+
   pageQuery.data.allContentstackPage.nodes.forEach(node => {
-    if (node.url !== "/" && node.url !== "/blog") {
-      createPageTemplate(node.url, pageTemplate, node.url)
+    if (node.url === "/blog") {
+      createBlogPageTemplate(node.url, blogPageTemplate, node.title, node);
     }
-  })
-}
+  });
+
+  pageQuery.data.allContentstackPage.nodes.forEach(node => {
+    if (node.url !== "/blog") {
+      createPageTemplate(node.url, pageTemplate, node.url);
+    }
+  });
+};
